Moved fully to static lexer
