+++ .github/workflows/rust.yml
@@ -15,6 +15,6 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Build
-      run: cargo build --verbose --all-features
+      run: cargo build --verbose
     - name: Run tests
-      run: cargo test --verbose --all-features
+      run: cargo test --verbose
