+++ src/compiling/mod.rs
@@ -5,14 +5,12 @@
 
 mod lexing;
 
-#[cfg(feature = "compiler")]
 mod compiler;
 mod token;
 mod parser;
 pub mod failing;
 
 pub use lexing::*;
-#[cfg(feature = "compiler")]
 pub use compiler::*;
 pub use token::*;
 pub use parser::*;
